feat: mitt优化，拆分types，新增addEvents事件绑定和watch监听事件
