+++ src/events/index.ts
@@ -1,1 +1,3 @@
 export * from './mitt';
+export * from './addEvent';
+export * from './watch';
