+++ src/events/mitt.ts
@@ -1,50 +1,15 @@
+import { EventType, EventHandlerMap, Emitter, WildcardHandler, Handler, EventHandlerList, WildCardEventHandlerList } from "./types"
 /**
  * @func mitt
  * @desc mitt 事件订阅发布及调度中心
  * @github https://github.com/developit/mitt
  * @example
- *    const emitter = mitt(); 
+ *    const emitter = mitt();
  *    emitter.on('foo', e => console.log('foo', e) )// 事件监听
  *    function onFoo() {}
  *    emitter.on('foo', onFoo)   // listen
  *    emitter.off('foo', onFoo)  // unlisten
- */
-export type EventType = string | symbol;
-// An event handler can take an optional event argument
-// and should not return a value
-export type Handler<T = unknown> = (event: T) => void;
-export type WildcardHandler<T = Record<string, unknown>> = (
-	type: keyof T,
-	event: T[keyof T]
-) => void;
-
-// An array of all currently registered event handlers for a type
-export type EventHandlerList<T = unknown> = Array<Handler<T>>;
-export type WildCardEventHandlerList<T = Record<string, unknown>> = Array<WildcardHandler<T>>;
-
-// A map of event types and their corresponding event handlers.
-export type EventHandlerMap<Events extends Record<EventType, unknown>> = Map<
-	keyof Events | '*',
-	EventHandlerList<Events[keyof Events]> | WildCardEventHandlerList<Events>
->;
-
-export interface Emitter<Events extends Record<EventType, unknown>> {
-	all: EventHandlerMap<Events>;
-
-	on<Key extends keyof Events>(type: Key, handler: Handler<Events[Key]>): void;
-	on(type: '*', handler: WildcardHandler<Events>): void;
-
-	off<Key extends keyof Events>(type: Key, handler?: Handler<Events[Key]>): void;
-	off(type: '*', handler: WildcardHandler<Events>): void;
-
-	emit<Key extends keyof Events>(type: Key, event: Events[Key]): void;
-	emit<Key extends keyof Events>(type: undefined extends Events[Key] ? Key : never): void;
-}
-
-/**
- * Mitt: Tiny (~200b) functional event emitter / pubsub.
- * @name mitt
- * @returns {Mitt}
+ * @returns { Mitt }
  */
 export default function mitt<Events extends Record<EventType, unknown>>(
 	all?: EventHandlerMap<Events>
